+++ tests/auto/corelib/serialization/qcborvalue/tst_qcborvalue.cpp
@@ -229,7 +229,7 @@ void tst_QCborValue::basics_data()
             if (t == QCborValue::Double)
                 return QTest::addRow("Double:%g", exp.toDouble());
             if (t == QCborValue::ByteArray || t == QCborValue::String)
-                return QTest::addRow("%s:%d", typeString, exp.toString().size());
+                return QTest::addRow("%s:%zd", typeString, size_t(exp.toString().size()));
             return QTest::newRow(typeString);
         };
         addRow() << t << v << exp;
